Move more common classes into the only library that uses them
Remove some unused constants from CommonConstants, remove some unnecessary client-side validation of device/module symmetric keys
