+++ iothub/service/src/Authentication/AuthenticationMethodFactory.cs
