+++ iothub/service/src/Authentication/IAuthorizationHeaderProvider.cs
