+++ iothub/service/src/Microsoft.Azure.Devices.csproj
@@ -54,35 +54,14 @@
     <Compile Include="$(Common)\Logging.Common.cs">
       <Link>Common\Logging.Common.cs</Link>
     </Compile>
-    <Compile Include="$(Common)\service\AuthenticationMethodFactory.cs">
-      <Link>Authentication\AuthenticationMethodFactory.cs</Link>
-    </Compile>
-    <Compile Include="$(Common)\service\CommonConstants.cs">
-      <Link>Common\CommonConstants.cs</Link>
-    </Compile>
-    <Compile Include="$(Common)\service\CommonResources.cs">
-      <Link>Common\CommonResources.cs</Link>
-    </Compile>
     <Compile Include="$(Common)\service\ExceptionHandlingHelper.cs">
       <Link>ExceptionHandlingHelper.cs</Link>
-    </Compile>
-    <Compile Include="$(Common)\service\ExceptionTrace.cs">
-      <Link>Common\ExceptionTrace.cs</Link>
-    </Compile>
-    <Compile Include="$(Common)\service\FatalException.cs">
-      <Link>Common\Exceptions\FatalException.cs</Link>
     </Compile>
     <Compile Include="$(Common)\service\FxCop.cs">
       <Link>Common\FxCop.cs</Link>
     </Compile>
-    <Compile Include="$(Common)\service\IAuthorizationHeaderProvider.cs">
-      <Link>Authentication\IAuthorizationHeaderProvider.cs</Link>
-    </Compile>
     <Compile Include="$(Common)\service\IotHubConnectionString.cs">
       <Link>Authentication\IotHubConnectionString.cs</Link>
-    </Compile>
-    <Compile Include="$(Common)\service\StringValidationHelper.cs">
-      <Link>Common\StringValidationHelper.cs</Link>
     </Compile>
     <Compile Include="$(Common)\service\TaskHelpers.cs">
       <Link>Common\TaskHelpers.cs</Link>
