+++ iothub/service/src/Utilities/CommonResources.cs
