+++ iothub/service/src/Utilities/ExceptionTrace.cs
