+++ iothub/service/src/Utilities/FatalException.cs
